# public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SCPM API",
        "description": "HTTP JSON API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/solidCP/enterprise-dispatchers": {
            "get": {
                "tags": [
                    "Enterprise Dispatchers"
                ],
                "description": "Get list of Enterprise Dispatchers, use only if you have more than one Enterprise Dispatchers",
                "operationId": "2642c66dd2d037f80efb023be8bbf9bd",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "return ids and names",
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SimpleError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/solidCP/enterprise-dispatchers/default/is-enable": {
            "get": {
                "tags": [
                    "Enterprise Dispatchers"
                ],
                "description": "Check if default Enterprise Dispatcher is manually disabled",
                "operationId": "c450b13cb11e800640c3c75e012b97a4",
                "parameters": [
                    {
                        "name": "id_enterprise_dispatcher",
                        "in": "query",
                        "description": "if not pass value, it will use a default",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isEnable": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SimpleError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorModel": {
                "properties": {
                    "error": {
                        "nullable": true,
                        "properties": {
                            "code": {
                                "type": "integer"
                            },
                            "message": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "violations": {
                        "nullable": true,
                        "type": "array",
                        "items": {
                            "properties": {
                                "propertyPath": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "SimpleError": {
                "properties": {
                    "error": {
                        "nullable": true,
                        "properties": {
                            "code": {
                                "type": "integer"
                            },
                            "message": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}